wip: working on activities filter hook
