+++ public/static/locales/en/validators.json
@@ -12,5 +12,7 @@
   "active": "Active",
   "inactive": "Inactive",
   "searchActiveValidator": "Search active validator",
-  "searchInactiveValidator": "Search inactive validator"
+  "searchInactiveValidator": "Search inactive validator",
+  "powerEvents": "Power Events",
+  "activities": "Activities"
 }
