+++ src/components/activities_filter/hooks.tsx
@@ -1,0 +1,3 @@
+export const useActivitiesFilter = () => {
+
+};
