+++ src/components/activities_filter/utils.tsx
@@ -1,0 +1,28 @@
+export const getFilterLabels = () => {
+  return ([
+    {
+      key: 'none',
+      display: 'none',
+    },
+    {
+      key: 'transfer',
+      display: 'transfer',
+    },
+    {
+      key: 'staking',
+      display: 'staking',
+    },
+    {
+      key: 'distribution',
+      display: 'distribution',
+    },
+    {
+      key: 'governance',
+      display: 'governance',
+    },
+    {
+      key: 'slashing',
+      display: 'slashing',
+    },
+  ]);
+};
