+++ src/screens/activities/utils.tsx
@@ -208,32 +208,3 @@ export const getCollapsibleLabels = (t:any) => {
     memo: t('memo'),
   });
 };
-
-export const getFilterLabels = () => {
-  return ([
-    {
-      key: 'none',
-      display: 'none',
-    },
-    {
-      key: 'transfer',
-      display: 'transfer',
-    },
-    {
-      key: 'staking',
-      display: 'staking',
-    },
-    {
-      key: 'distribution',
-      display: 'distribution',
-    },
-    {
-      key: 'governance',
-      display: 'governance',
-    },
-    {
-      key: 'slashing',
-      display: 'slashing',
-    },
-  ]);
-};
