+++ src/screens/validator_details/components/index.ts
@@ -1,5 +1,7 @@
 import Info from './info';
+import PowerActivities from './power_activities';
 
 export {
   Info,
+  PowerActivities,
 };
