+++ src/screens/validator_details/components/power_activities/hooks.tsx
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+
+export const useValidatorDetailsHook = () => {
+  const [tabValue, setTabValue] = useState(0);
+
+  const handleTabChange = (_event:any, newValue: number) => {
+    setTabValue(newValue);
+  };
+
+  return {
+    tabValue,
+    handleTabChange,
+  };
+};
